Improve compatibility with WARN_CREATE_GLOBAL
In the interest of improving compatibility with the zsh option
`WARN_CREATE_GLOBAL`, include several more declarations of variables,
to explicitly declare whether they are to be local or global.

Without this change, several variables are created as global from
within functions, causing warnings with `WARN_CREATE_GLOBAL`. In some
cases, this would appear to be the intended behavior; in most cases,
it would not.